proof of concept: spaceauth now works
after discovery, a new challenge is generated and written to the challenge attribute. When finished calculating, the coin sends an indication: the `notify_func` callback is called. It requests the whole response. In `read_func`, the response is written to the appropriate array. When the last bytes are written, it verifies the response and clears challenge and response. After verification, the connection is closed as it isn't needed any further.